refactor: simplify diff
